+++ src/main/java/io/github/computerdaddyguy/jfiletreeprettyprinter/DefaultFileTreePrettyPrinter.java
@@ -4,7 +4,9 @@ import io.github.computerdaddyguy.jfiletreeprettyprinter.renderer.TreeEntryRende
 import io.github.computerdaddyguy.jfiletreeprettyprinter.scanner.PathToTreeScanner;
 import java.nio.file.Path;
 import java.util.Objects;
+import java.util.function.Predicate;
 import org.jspecify.annotations.NullMarked;
+import org.jspecify.annotations.Nullable;
 
 @NullMarked
 class DefaultFileTreePrettyPrinter implements FileTreePrettyPrinter {
@@ -18,8 +20,8 @@ class DefaultFileTreePrettyPrinter implements FileTreePrettyPrinter {
 	}
 
 	@Override
-	public String prettyPrint(Path path) {
-		var tree = scanner.scan(path);
+	public String prettyPrint(Path path, @Nullable Predicate<Path> filter) {
+		var tree = scanner.scan(path, filter);
 		return renderer.renderTree(tree);
 	}